move the umap map from program to venue page
I think this was added to the program details during the conference
to help people find the venues for the different things on the
program. But at this this stage people will go to the program page
primarily to find the details and videos of the sessions and so at
this stage that map is rather cluttering the page.

The map does however fit quite nicely with the transport info etc
on the venue page